+++ test/fixtures/mod-private-proxy/app/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"git.fury.io/go-test-org/gotest"
+)
+
+func main() {
+	gotest.PrintHello()
+}