+++ test/fixtures/mod-with-hg-dep/main.go
@@ -1,9 +1,0 @@
-package main
-
-import "fmt"
-import "bitbucket.org/pkg/inflect"
-
-func main() {
-	i := inflect.Rule{}
-	fmt.Println(i)
-}